use sdk assert in examples
