+++ examples/package.json
@@ -33,11 +33,11 @@
   "license": "Apache-2.0",
   "dependencies": {
     "lodash-es": "^4.17.21",
-    "near-sdk-js": "file:../"
+    "near-sdk-js": "file:../near-sdk-js-v0.4.0-2.tgz"
   },
   "devDependencies": {
     "ava": "^4.2.0",
-    "typescript": "^4.7.4",
-    "near-workspaces": "2.0.0"
+    "near-workspaces": "2.0.0",
+    "typescript": "^4.7.4"
   }
 }
