+++ lib/api.js
@@ -289,3 +289,6 @@ export function storageRemove(key) {
     }
     return false;
 }
+export function storage_byte_cost() {
+    return 10000000000000000000n;
+}
