+++ lib/index.d.ts
@@ -2,6 +2,5 @@ import { call, view, NearBindgen } from "./near-bindgen";
 import { NearContract } from "./near-contract";
 import * as near from "./api";
 import { LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet } from "./collections";
-import { bytes, Bytes } from "./utils";
-import * as utils from "./utils";
-export { call, view, NearBindgen, NearContract, near, LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet, bytes, Bytes, utils, };
+import { bytes, Bytes, assert } from "./utils";
+export { call, view, NearBindgen, NearContract, near, LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet, bytes, Bytes, assert, };
