+++ lib/index.js
@@ -2,6 +2,5 @@ import { call, view, NearBindgen } from "./near-bindgen";
 import { NearContract } from "./near-contract";
 import * as near from "./api";
 import { LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet, } from "./collections";
-import { bytes } from "./utils";
-import * as utils from "./utils";
-export { call, view, NearBindgen, NearContract, near, LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet, bytes, utils, };
+import { bytes, assert } from "./utils";
+export { call, view, NearBindgen, NearContract, near, LookupMap, Vector, LookupSet, UnorderedMap, UnorderedSet, bytes, assert, };
