+++ lib/utils.d.ts
@@ -2,5 +2,4 @@ export declare type Bytes = string;
 export declare function u8ArrayToBytes(array: Uint8Array): string;
 export declare function bytesToU8Array(bytes: Bytes): Uint8Array;
 export declare function bytes(strOrU8Array: string | Uint8Array): Bytes;
-export declare function storage_byte_cost(): BigInt;
 export declare function assert(b: boolean, str: string): void;
