+++ lib/utils.js
@@ -31,9 +31,6 @@ function checkStringIsBytes(str) {
     }
     return str;
 }
-export function storage_byte_cost() {
-    return 10000000000000000000n;
-}
 export function assert(b, str) {
     if (b) {
         return;
