+++ src/utils.ts
@@ -41,10 +41,6 @@ function checkStringIsBytes(str: string) {
   return str;
 }
 
-export function storage_byte_cost(): BigInt {
-  return 10_000_000_000_000_000_000n;
-}
-
 export function assert(b: boolean, str: string) {
   if (b) {
       return
